% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehashMAE.R
\name{dehashMAE}
\alias{dehashMAE}
\title{deanonymize a rehashed MultiAssayExperiment and its constitutive objects}
\usage{
dehashMAE(x, meta = NULL, covs = NULL, check = FALSE)
}
\arguments{
\item{x}{a MultiAssayExperiment to deanonymize}

\item{meta}{the meta-metadata required to reverse the hashes}

\item{covs}{a data.frame of covariates (rows are samples, columns covs)}

\item{check}{check assay row and column hashes? (FALSE; can be very slow)}
}
\value{
\preformatted{     a dehashed MultiAssayExperiment
}
}
\description{
Warning: this is not implemented properly yet.
}
