% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehashSE.R
\name{dehashSE}
\alias{dehashSE}
\alias{dehashSummarizedExperiment}
\alias{rehydrateSummarizedExperiment}
\alias{rehydrateSE}
\title{deanonymize (to some degree) a SummarizedExperiment post public dissemination}
\usage{
dehashSE(x, meta = NULL, covs = NULL, check = FALSE)
}
\arguments{
\item{x}{a \link{Ranged}SummarizedExperiment to deanonymize}

\item{meta}{the meta-metadata required to reverse the hashes}

\item{covs}{a data.frame of covariates (rows are samples, columns covs)}

\item{check}{check assay row and column hashes? (FALSE; can be very slow)}
}
\value{
\preformatted{   an object of the same class as x, but with cleartext dimnames
}
}
\description{
Verifiable de-anonymization requires metadata: a sample map, a feature map,
a salting phrase, and knowledge of whether the salt phrase was applied to
features, samples, both, or neither. If the metadata is left in the metadata
slot of the dehydrated SummarizedExperiment, it will be used, else it must
be supplied by the user. Optionally, the data itself may be verified.
}
