% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehashURL.R
\name{dehashURL}
\alias{dehashURL}
\title{dehash (if necessary) the file at an URL. Note: this function is NOT 'safe'!}
\usage{
dehashURL(x, ...)
}
\description{
Seriously -- if you aren't sure that you trust the owner of an URL and the
payload, you MUST NOT apply this function to it. We'll try to snoop the file
header, but that is NOT strong protection. Download and inspect if unsure.
}
