% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssayColHashes.R
\name{getAssayColHashes}
\alias{getAssayColHashes}
\title{get column-wise hashes for each assay in a SummarizedExperiment-like object}
\usage{
getAssayColHashes(x, algo = "md5")
}
\arguments{
\item{x}{a SummarizedExperiment-like object}

\item{algo}{which algorithm to use (default "md5")}
}
\value{
\preformatted{ a data.frame with column-wise hashes, one assay per column
}
}
\description{
get column-wise hashes for each assay in a SummarizedExperiment-like object
}
