% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssayRowHashes.R
\name{getAssayRowHashes}
\alias{getAssayRowHashes}
\title{get row-wise digests for each assay in a SummarizedExperiment as a dataframe}
\usage{
getAssayRowHashes(x, algo = "md5")
}
\arguments{
\item{x}{a SummarizedExperiment-like object}

\item{algo}{which algorithm to use (default "md5")}
}
\value{
\preformatted{ a data.frame of row-wise assay hashes, one column per assay 
}
}
\description{
get row-wise digests for each assay in a SummarizedExperiment as a dataframe
}
