% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTPM.R
\name{getTPM}
\alias{getTPM}
\title{Get transcripts per million (TPM) from a recount RangedSummarizedExperiment}
\usage{
getTPM(object, ...)
}
\arguments{
\item{object}{an RSE from recount}

\item{...}{any additional arguments to pass to recount::getRPKM}
}
\value{
\preformatted{   a matrix of TPM estimates
}
}
\description{
Get transcripts per million (TPM) from a recount RangedSummarizedExperiment
}
\examples{
library(recount)
getTPM(rse_gene_SRP009615)

}
\seealso{
download_study

rse_gene_SRP009615
}
