% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rehashMAE.R
\name{rehashMAE}
\alias{rehashMAE}
\title{anonymize (to some degree) a MultiAssayExperiment and its constituent objects}
\usage{
rehashMAE(x, salt = "0x", strip = TRUE, algo = "md5",
  deorder = FALSE)
}
\arguments{
\item{x}{a MultiAssayExperiment to anonymize}

\item{salt}{a salting phrase to slow brute-force attacks ("0x")}

\item{strip}{strip rehashed objects of any deID'ing metadata? (TRUE)}

\item{algo}{algorithm to use for the one-way hash (default is "md5")}

\item{deorder}{scramble rows and columns? (FALSE; disrupts data digest)}
}
\value{
\preformatted{         a new, rehashed MultiAssayExperiment
}
}
\description{
Warning: this function is not yet properly implemented.
}
