% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rehashURL.R
\name{rehashURL}
\alias{rehashURL}
\title{take a hashed (or not, doesn't matter) URL and dehash (if needed) the .rda}
\usage{
rehashURL(object, ...)
}
\description{
take a hashed (or not, doesn't matter) URL and dehash (if needed) the .rda
}
