% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revertGrSet.R
\name{revertGrSet}
\alias{revertGrSet}
\title{turn an HDF5-backed SummarizedExperiment back into a usable GenomicRatioSet}
\usage{
revertGrSet(grSet, HM = c("EPIC", "HM450", "HM27"))
}
\arguments{
\item{grSet}{a SummarizedExperiment, often HDF5-backed, from a grSet}

\item{HM}{optional indicator of what platform grSet is made from}
}
\description{
This function depends on the following to figure out how to annotate:
}
\details{
minfi:::.default.27k.annotation
minfi:::.default.450k.annotation
minfi:::.default.epic.annotation

If there's a disconnect between these and the truth, it will cause problems.
}
