% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPKMtoTPM.R
\name{RPKMtoTPM}
\alias{RPKMtoTPM}
\title{Reads per kilobase per million (RPKM) to transcripts per million (TPM)}
\usage{
RPKMtoTPM(object, ...)
}
\arguments{
\item{object}{a (DelayedM|m)atrix of RPKM/FPKMs, or something containing one}

\item{...}{any additional user input arguments to pass to other functions}
}
\value{
\preformatted{   a (DelayedM|m)atrix of TPM estimates, or an object with one.
}
}
\description{
Per Colin Dewey, TPM == (RPKM\link{j} / sum(RPKM)) * 1000000, for every feature j.
(The same is true of FPKM, since RPKM just estimates fragments of template.)
}
\examples{
library(recount)
RPKMtoTPM(getRPKM(rse_gene_SRP009615))

}
\seealso{
http://bioinformatics.oxfordjournals.org/content/26/4/493.full
}
