% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripDimnames.R
\name{stripDimnames}
\alias{stripDimnames}
\title{strip dimnames (forcibly) for verifiable row/column/assay hash checking}
\usage{
stripDimnames(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
a matrix with no dimnames
}
\description{
strip dimnames (forcibly) for verifiable row/column/assay hash checking
}
