% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehash.R
\name{dehash}
\alias{dehash}
\title{a virtual function, since one-way hashes only make sense forwards... thus
delegates (dehashSE for SummarizedExperiments, dehashSCE for
SingleCellExperiments, dehashBSseq for BSseq objects, and so forth) do all
of the heavy lifting for this method.}
\usage{
dehash(x, ...)
}
\arguments{
\item{x}{the thing to dehash}

\item{...}{arguments to pass to the called function for objects like x}
}
\value{
\preformatted{ an object of the same class as x, but with identities restored
}
}
\description{
a virtual function, since one-way hashes only make sense forwards... thus
delegates (dehashSE for SummarizedExperiments, dehashSCE for
SingleCellExperiments, dehashBSseq for BSseq objects, and so forth) do all
of the heavy lifting for this method.
}
\seealso{
rehash

dehashSE

dehashSCE

dehashBSseq
}
