% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehashSE.R
\name{dehashSE}
\alias{dehashSE}
\alias{dehashSummarizedExperiment}
\alias{rehydrateSummarizedExperiment}
\alias{rehydrateSE}
\title{deanonymize (to some degree) a SummarizedExperiment post public dissemination}
\usage{
dehashSE(x, meta = NULL, covs = NULL, check = FALSE)
}
\arguments{
\item{x}{a \link{Ranged}SummarizedExperiment to deanonymize, or a list*}

\item{meta}{the meta-metadata required to reverse the hashes}

\item{covs}{a data.frame of covariates (rows are samples, columns covs)}

\item{check}{check assay row and column hashes? (FALSE; can be very slow)}
}
\value{
\preformatted{   an object of the same class as x, but with cleartext dimnames
}
}
\description{
Verifiable de-anonymization requires metadata: a sample map, a feature map,
a salting phrase, and knowledge of whether the salt phrase was applied to
features, samples, both, or neither. If the metadata is left in the metadata
slot of the dehydrated SummarizedExperiment, it will be used, else it must
be supplied by the user. Optionally, the data itself may be verified.
}
\details{
\itemize{
\item A list result from rehash(x, ..., strip=TRUE) can be passed as \code{x} (the
first argument to dehashSE) for testing purposes. In such cases, it will
be disassembled into the necessary symbols to dehash the object. However,
saving the list for distribution defeats the purpose of rehash()ing it,
so don't do that in production. This hook exists ONLY for testing! If a
list is detected as the argument, row and column hashes WILL be checked
(regardless of the value of \code{check} in the function call), partly to
discourage routine use of this functionality in production.
}
}
