% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rehash.R
\name{rehash}
\alias{rehash}
\title{a salty hashing function (and a generic that dispatches it)}
\usage{
rehash(x, salt = NULL, ...)
}
\arguments{
\item{x}{the string to hash}

\item{salt}{the salt to apply (NULL)}

\item{...}{other parameters to pass on to digest::digest}
}
\description{
a salty hashing function (and a generic that dispatches it)
}
